Working on separating out the parameters modal.
