+++ src/main/resources/i18n/messages.properties
@@ -1603,6 +1603,8 @@ LaunchDetails.name.required=Workflow Pipeline require a name for easy identifica
 LaunchDetails.description=Description
 
 SavedParameters.title=Parameters
+SavedParameters.useModified.btn=Use Modified Parameters
+SavedParameters.useModified.tooltip=Parameters will be used by the current pipeline only and will not be saved.
 
 # ========================================================================================== #
 # PIPELINES - Phylogenomics                                                                  #
