+++ src/main/webapp/resources/js/pages/launch/components/ParametersModal/ParametersModal.jsx
@@ -1,19 +1,9 @@
 import React from "react";
-import {
-  Alert,
-  Button,
-  Dropdown,
-  Form,
-  Input,
-  Menu,
-  Modal,
-  Space,
-  Tooltip,
-} from "antd";
+import isEqual from "lodash/isEqual";
+import { Form, Input, Modal } from "antd";
 import { useLaunchDispatch, useLaunchState } from "../../launch-context";
-import { IconExclamationCircle } from "../../../../components/icons/Icons";
-import { SPACE_SM, SPACE_XS } from "../../../../styles/spacing";
-import { grey9, yellow6 } from "../../../../styles/colors";
+import { SPACE_SM } from "../../../../styles/spacing";
+import { ParametersFooter } from "./ParametersFooter";
 
 /**
  * React component to render a modal window for modifying and saving pipeline
@@ -31,17 +21,36 @@ export function ParametersModal({ visible, closeModal }) {
    parameter set.
    */
   const { parameterSet } = useLaunchState();
+
   const {
     dispatchUseModifiedParameters,
     dispatchOverwriteParameterSave,
   } = useLaunchDispatch();
 
+  /*
+  Store a copy of the original values to compare against to see if they
+  have been modified.
+   */
+  const originalValues = parameterSet.parameters.reduce(
+    (a, b) => ({ ...a, [b.name]: b.value }),
+    {}
+  );
+
+  /*
+  Have any of the field been modified?
+   */
+  const [modified, setModified] = React.useState(false);
+
+  /*
+  Use by the ant form to set original values for the parameters.
+   */
   const [fields, setFields] = React.useState([]);
-  const [showSave, setShowSave] = React.useState(false);
-  const [showSaveAs, setShowSaveAs] = React.useState(false);
 
   const [form] = Form.useForm();
 
+  /*
+  Set the original value for all fields on the form
+   */
   React.useEffect(() => {
     setFields(
       parameterSet.parameters.map((parameter) => ({
@@ -53,13 +62,6 @@ export function ParametersModal({ visible, closeModal }) {
 
   const saveParameters = (name) => {
     form.validateFields().then((values) => console.log(values));
-  };
-
-  const onOk = () => {
-    form.validateFields().then((values) => {
-      console.log(values);
-      closeModal();
-    });
   };
 
   const useModifiedParameters = () => {
@@ -84,18 +86,15 @@ export function ParametersModal({ visible, closeModal }) {
     form.validateFields().then(dispatchOverwriteParameterSave);
   };
 
-  const menu = (
-    <Menu>
-      {parameterSet.id !== 0 ? (
-        <Menu.Item key="copy" onClick={() => setShowSave(true)}>
-          Save
-        </Menu.Item>
-      ) : null}
-      <Menu.Item key="save" onClick={() => setShowSaveAs(true)}>
-        Save as ...
-      </Menu.Item>
-    </Menu>
-  );
+  /**
+   * Helper function to determine if the original parameters of this set have
+   * been modified (which would result in updates to the UI).
+   *
+   * @param {object} changedValues - most recent changed value
+   * @param {object} allValues - all current values in the form.
+   */
+  const onValuesChange = (changedValues, allValues) =>
+    setModified(!isEqual(originalValues, allValues));
 
   return (
     <Modal
@@ -106,72 +105,23 @@ export function ParametersModal({ visible, closeModal }) {
       keyboard={false}
       maskClosable={false}
       onCancel={closeModal}
-      onOk={onOk}
-      title={parameterSet.label}
+      title={`${parameterSet.label}${modified ? "*" : ""}`}
       width={600}
       footer={
-        <Space style={{ width: `100%` }} direction="vertical">
-          {showSaveAs || showSave ? null : (
-            <div style={{ display: "flex", flexDirection: "row-reverse" }}>
-              <Space>
-                <Button onClick={closeModal}>Cancel</Button>
-                <Dropdown.Button
-                  type="primary"
-                  overlay={menu}
-                  buttonsRender={([leftButton, rightButton]) => [
-                    <Tooltip
-                      color={yellow6}
-                      title={
-                        <span style={{ color: grey9 }}>
-                          <IconExclamationCircle
-                            style={{ marginRight: SPACE_XS }}
-                          />
-                          Parameters will not be saved
-                        </span>
-                      }
-                      key="leftButton"
-                    >
-                      {React.cloneElement(leftButton, {
-                        onClick: useModifiedParameters,
-                      })}
-                    </Tooltip>,
-                    rightButton,
-                  ]}
-                >
-                  <div>Use these parameters</div>
-                </Dropdown.Button>
-              </Space>
-            </div>
-          )}
-          {showSaveAs ? (
-            <Form layout="vertical">
-              <Form.Item label={"NAME"}>
-                <Input defaultValue={`${parameterSet.label} (copy)`} />
-              </Form.Item>
-              <Button onClick={() => setShowSaveAs(false)}>Cancel</Button>
-              <Button onClick={() => {}}>Save</Button>
-            </Form>
-          ) : null}
-          {showSave ? (
-            <Form layout="vertical">
-              <Alert
-                style={{ textAlign: "left", marginBottom: SPACE_SM }}
-                type="warning"
-                showIcon
-                message="This will overwrite the current parameter set"
-                description={
-                  "EVerytihng else will magically be the same. Have fun and do evil things."
-                }
-              />
-              <Button onClick={() => setShowSave(false)}>Cancel</Button>
-              <Button onClick={() => overwriteParameterSet()}>Save</Button>
-            </Form>
-          ) : null}
-        </Space>
+        <ParametersFooter
+          modified={modified}
+          onCancel={closeModal}
+          useModifiedParameters={useModifiedParameters}
+        />
       }
     >
       <section style={{ maxHeight: 600, overflow: "auto", padding: SPACE_SM }}>
-        <Form form={form} layout="vertical" fields={fields}>
+        <Form
+          form={form}
+          layout="vertical"
+          fields={fields}
+          onValuesChange={onValuesChange}
+        >
           {parameterSet.parameters.map((parameter) => (
             <Form.Item
               key={parameter.name}
